refactor: enum 클래스 패키지명 변경 (#CHZZ-80) - enums -> type 패키지명 변경
