+++ src/test/java/org/chzz/market/domain/auction/repository/AuctionRepositoryImplTest.java
@@ -1,7 +1,7 @@
 package org.chzz.market.domain.auction.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.chzz.market.domain.auction.enums.AuctionStatus.*;
+import static org.chzz.market.domain.auction.type.AuctionStatus.*;
 
 import java.time.LocalDateTime;
 import java.util.Comparator;
